Add support for Alibaba Cloud Linux
The installation process for Tailscale on Alibaba Cloud Linux 3 is
similar to the one used for Amazon Linux 2023. Unfortunately newer
versions of Ansible identify "yum" as the package manager for this OS,
which makes the `ansible.builtin.dnf` module request a `use_backend`
parameter. Using a newer version of ansible doesn't fix the problem
either for [unrelated
reasons](https://www.jeffgeerling.com/blog/2024/newer-versions-ansible-dont-work-rhel-8).

I've created different install/uninstall steps specific to Alibaba Cloud
Linux 3, which are almost identical to the "CentOS and related families"
tasks. Difference being the `use_backend` param for the dnf module.

Tailscale doesn't provide specific repo packages for Alibaba Cloud
Linux, so the repo path is the same as CentOS 8, which what this OS is
based on.